Cereebro Server starters : Standalone and Eureka server add-on
Includes several refactorings, too big of a commit actually.

`cereebro-server` is now somewhat of an autoconfigure module, 
but it also pulls transitive dependencies like webjars.  
It contains an `@EnableCereebroServer` annotation that is 
required to configure a cereebro server, even if the 
`cereebro-server-spring-boot-starter` is on the classpath.

`cereebro-server-eureka-addon-spring-cloud-starter` is to be 
used in a Eureka Server spring cloud application.  
Instead of having two separate servers, cereebro can be used 
as an add-on for the Eureka server, fetching snitch metadata 
directly from the internal registry.

Images have been removed from the cereebro server.  
Absolute URLs caused issues when a context-path was defined.

Sample/test applications have been moved to sample projects.

CereebroEndpoint has been renamed CereebroSnitchEndpoint to 
prevent any conflict with the CereebroSystemController. 
Both use different URLs now : 

 * `/cereebro/snitch` for the cereebro snitch endpoint
 * `/cereebro/system` for the cereebro server resources

Closes #41
Closes #42