Add file writer snitch, ApplicationAnalyzer and refactor
Write to a file on startup, depending on config props. 
Temporary folder is used when activated without a 
location.

`ApplicationAnalyzer` is an "internal snitch" within 
ann application, it does not have a URL but can give 
a `SystemFragment`. Created so other `Snitch` 
implementations can reuse it.

Move around Slf4jLogSnitch.

Refactor exceptions in unit tests.

Closes #52