Improve UI, handle table differently
Remove "arrow" type edges, keep the graph looking like a graph. 
Create a graph tab in navbar, preparing the diagram view.
Component list is either a table or a list of divs 
depending on the viewport.
Add utility methods on core classes.

Add styling and gimmicks like a navbar and footer.
Would be a lot cooler with a logo.

Closes #56