Change the way Eureka metadata is registered
Because Zuul relationship detector polled the Eureka 
Client to find downstream services, application was
forced to register before we could populate Cereebro 
metadata.  

This commit makes `EurekaMetadataPopulator` force the 
registration of metadata _after_ we have detected 
relationships.
We use `ApplicationInfoManager` instead of `EurekaInstanceConfigBean` in
order to have access to the method that forces metadata registration.

Fixes #81